+++ public/app/plugins/datasource/cloud-monitoring/functions.test.ts
@@ -9,6 +9,7 @@ import {
   getMetricTypesByService,
   labelsToGroupedOptions,
   stringArrayToFilters,
+  alignmentPeriodLabel,
 } from './functions';
 import { newMockDatasource } from './specs/testData';
 import { AlignmentTypes, MetricDescriptor, MetricKind, ValueTypes } from './types';
@@ -224,6 +225,15 @@ describe('functions', () => {
           value: 'value',
         },
       ]);
+    });
+  });
+
+  describe('alignmentPeriodLabel', () => {
+    it('returns period label if alignment period and per series aligner is set', () => {
+      const datasource = newMockDatasource();
+
+      const label = alignmentPeriodLabel({ perSeriesAligner: 'ALIGN_DELTA', alignmentPeriod: '10' }, datasource);
+      expect(label).toBe('10s interval (delta)');
     });
   });
 });
